changed items to list shop items replaced by myitems command
