+++ commands/items.js
@@ -1,20 +1,20 @@
 const { SlashCommandBuilder } = require('discord.js');
 const fs = require("fs");
 
+var items;
+fs.readFile('./data/item_data.json', (err, data) => {
+    if (err) throw err;
+    items = JSON.parse(data);
+});
+
 module.exports = {
     data: new SlashCommandBuilder()
         .setName("items")
-        .setDescription("Lists the user's items"),
+        .setDescription("Lists the items available to buy"),
         async execute(interaction) {
-            let server = interaction.guild;
-            let author = interaction.user;
-            fs.readFile('./data/player_item_data.json', (err, data) => {
-                if (err) throw err;
-                let item_data = JSON.parse(data);
-                for (let item in item_data[server.id][author.id]) {
-                    interaction.channel.send(`${item}: ${item_data[server.id][author.id][item]}`);
-                }
-            });
-            interaction.reply("Your items: ");
+            for (let item in items) {
+                interaction.channel.send(`${item}: $${items[item]["price"]}`);
+            }
+            interaction.reply("Store: ");
         }
-}
+    }
